requirements.txt
Flask==2.3.3
gunicorn==21.2.0
torch
torchvision
timm==0.9.12
facenet-pytorch==2.5.3
numpy==1.26.4
opencv-python-headless==4.8.1.78
Pillow==10.0.1